Split out state machine; instructions -> actions
